requirements.txt
arxiv
beautifulsoup4
bs4
chromadb
duckdb
duckduckgo-search
faiss-cpu
fastapi
gradio
huggingface_hub
ipykernel
langchain
langchain-community
langchain-openai
langchain-text-splitters
langchain-chroma
langchain-classic
langchain-groq
langchain-huggingface
langserve
mysql-connector-python
numexpr
openai
pandas
pypdf
pymupdf
python-dotenv
pytube
sentence-transformers
sse-starlette
streamlit
SQLAlchemy
unstructured
uvicorn
validators==0.28.1
wikipedia
youtube_transcript_api
langchain-yt-dlp
ddgs